Remove duplicate secure_primitives! macro
