+++ crates/holochain_integrity_types/Cargo.toml
@@ -23,8 +23,9 @@ kitsune_p2p_timestamp = { version = "^0.0.15", path = "../kitsune_p2p/timestamp"
 subtle = "2"
 
 # Optional
-tracing = { version = "0.1", optional = true }
 arbitrary = { version = "1.0", features = ["derive"], optional = true }
+subtle-encoding = {version = "0.5", optional = true}
+tracing = { version = "0.1", optional = true }
 
 [features]
 default = []
